rodi_robot: Reduce the rospy rate to 1 Hz
The rospy rate for the main loop is 2 Hz, but RoDI sometimes can't cope
with that since the sonar values are queried on each iteration, causing
too much stress to RoDI and making a lot of HTTP connections to timeout.

Change to 1 Hz instead which should give a sonar lecture every 1 second.
That should be enough for most needs and reduces the number of timeouts.